test: split client tests between async and sync directory
